`bug security`

+++ DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/hp_29_MiniatureCrafts/dto/SupplierDTO.java
@@ -5,7 +5,7 @@ import com.example.hp_29_MiniatureCrafts.entity.Supplier;
 import java.time.LocalDate;
 
 public class SupplierDTO {
-    private Integer id;
+    private Long id;
     private String name;
     private String phone;
     private String address;
@@ -16,11 +16,11 @@ public class SupplierDTO {
 
     // Getters and Setters
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -94,7 +94,7 @@ public class SupplierDTO {
         this.editDate = supplier.getEdit_Date();
     }
 
-    public SupplierDTO(Integer id, String name, String phone, String address, String note, Boolean status, LocalDate creationDate, LocalDate editDate) {
+    public SupplierDTO(Long id, String name, String phone, String address, String note, Boolean status, LocalDate creationDate, LocalDate editDate) {
         this.id = id;
         this.name = name;
         this.phone = phone;
